Move Disqus components to their own folder
