+++ www/src/js/views/modules/ModulePageContent.jsx
@@ -14,8 +14,8 @@ import { BULLET, scrollToHash } from 'utils/react';
 import { NAVTAB_HEIGHT } from 'views/layout/Navtabs';
 import ModuleTree from 'views/modules/ModuleTree';
 import LinkModuleCodes from 'views/components/LinkModuleCodes';
-import CommentCount from 'views/components/CommentCount';
-import DisqusComments from 'views/components/DisqusComments';
+import CommentCount from 'views/components/disqus/CommentCount';
+import DisqusComments from 'views/components/disqus/DisqusComments';
 import Online from 'views/components/Online';
 import Warning from 'views/errors/Warning';
 import SideMenu from 'views/components/SideMenu';
